# Improve performance of mesecon.turnon and mesecon.turnoff