### Improvements:
- Use a better fifo queue for `frontiers`. (Old one is still included as separate implementation for testing, remove before merging!)
  (Related: https://github.com/minetest/minetest/pull/7683 @HybridDog)
- Nodes that we certainly don't need to visit again (on-/off-turned conductors and nodes that are ignored (eg. air)) are marked as such via `pos_can_be_skipped` (a hashset using the hashed node pos), and are not visited again.
  This affects the `depth`, but what is that anyway?
- `mesecon.turnoff` had called `mesecon.is_receptor_on` in a loop for no reason. It was probably a copy-error.

(There's probably much more that could be done.)

### How to test:
Use worldedit to make a big conductor block. And turn it on and off.
For time measuring I've used globalstep dtime:
```lua
local function printc(msg)
	print(msg)
	minetest.chat_send_all(msg)
end

minetest.register_globalstep(function(dt)
	if dt < 0.1 then
		return
	end
	printc("globalstep dt: "..dt)
end)
```

The conductor cuboid I've used `has a volume of 45000 nodes (30*30*50)` and consists of mese and `wire_11111111_off`:

![screenshot_20210309_233911](https://user-images.githubusercontent.com/7613443/110552088-f85d1900-8136-11eb-803a-1f667c8d6068.png)


### Test results:
(My cpu was scaled to 1600MHz, I'm using luajit.)
branch | turnon | turnoff
---|---|---
this PR | ca. 1.3 s | ca. 1.6 s
master | ca. 12 s | ca. 14 s